+++ src/server.lisp
@@ -3,6 +3,8 @@
         #:f-underscore)
   ;; to load js dependencies after app was started
   (:import-from #:weblocks/app-dependencies)
+  (:import-from #:weblocks/app
+                #:with-app)
   ;; we need to depend on this package, because
   ;; lack:builder will try to find `LACK.MIDDLEWARE.SESSION`
   ;; package
@@ -12,7 +14,8 @@
   (:import-from #:weblocks/js/jquery)
  
   (:import-from #:weblocks/session
-                #:*session*)
+                #:make-session-middleware
+                #:with-session)
   (:import-from #:weblocks/hooks
                 #:prepare-hooks)
   (:import-from #:weblocks/routes
@@ -63,6 +66,10 @@
   object. Otherwise, nil.")
 
 
+(defvar *clack-output* nil
+  "Here we'll store all output from the Clack, because we don't want it to mix with weblocks own output.")
+
+
 (defclass server ()
   ((port :type integer
          :initarg :port
@@ -104,48 +111,57 @@ Make instance, then start it with ``start`` method."
           app)))))
 
 
+(defun make-response-for-clack (response)
+  (etypecase response
+    (weblocks/response:response
+     (list (get-code response)
+           (get-headers response)
+           ;; Here we use catch to allow to abort usual response
+           ;; processing and to return data immediately
+           (list (get-content response))))
+    (list response)
+    (function response)))
+
+
 (defmethod handle-http-request ((server server) env)
   "Weblocks HTTP dispatcher.
 This function serves all started applications and their static files."
 
-  (let* ((*session* (getf env :lack.session))
-         ;; This "hack" is needed to allow widgets to change *random-state*
-         ;; and don't interfere with other threads and requests
-         (*random-state* *random-state*))
-    (with-request ((make-request env))
-      ;; Dynamic hook :handle-http-request makes possible to write
-      ;; some sort of middlewares, which change *request* and *session*
-      ;; variables.
-      (prepare-hooks
-        (weblocks/hooks:with-handle-http-request-hook ()
+  (let (;; This "hack" is needed to allow widgets to change *random-state*
+        ;; and don't interfere with other threads and requests
+        (*random-state* *random-state*))
+    (with-session (env)
+      (with-request ((make-request env))
+        ;; Dynamic hook :handle-http-request makes possible to write
+        ;; some sort of middlewares, which change *request* and *session*
+        ;; variables.
+        (prepare-hooks
+          (weblocks/hooks:with-handle-http-request-hook ()
 
-          (let* ((path-info (getf env :path-info))
-                 (hostname (getf env :server-name))
-                 (route (get-route path-info))
-                 (app (search-app-for-request-handling path-info hostname)))
+            (let* ((path-info (getf env :path-info))
+                   (hostname (getf env :server-name))
+                   (route (get-route path-info))
+                   (app (search-app-for-request-handling path-info hostname)))
 
-            (log:debug "Processing request to" path-info)
+              (log:debug "Processing request to" path-info)
 
-            ;; If dependency found, then return it's content along with content-type
-            (cond
-              (route
-               (log:debug "Route was found" route)
-               (weblocks/routes:serve route env))
-              (app
-               (log:debug "App was found" route)
-               (let* ((response (handle-request app)))
-                 (list (get-code response)
-                       (get-headers response)
-                       ;; Here we use catch to allow to abort usual response
-                       ;; processing and to return data immediately
-                       (list (get-content response)))))
-              (t
-               (log:error "Application dispatch failed for" path-info)
+              ;; If dependency found, then return it's content along with content-type
+              (with-app app
+                (make-response-for-clack
+                 (cond
+                   (route
+                    (log:debug "Route was found" route)
+                    (weblocks/routes:serve route env))
+                   (app
+                    (log:debug "App was found" route)
+                    (handle-request app))
+                   (t
+                    (log:error "Application dispatch failed for" path-info)
 
-               (list 404
-                     (list :content-type "text/html")
-                     (list (format nil "File \"~A\" was not found.~%"
-                                   path-info)))))))))))
+                    (list 404
+                          (list :content-type "text/html")
+                          (list (format nil "File \"~A\" was not found.~%"
+                                        path-info))))))))))))))
 
 
 (defun start-server (server &key debug)
@@ -153,38 +169,47 @@ This function serves all started applications and their static files."
 
 If server is already started, then logs a warning and does nothing."
   
-  (if (get-handler server)
-      (log:warn "Webserver already started")
-      
-      ;; Otherwise, starting a server
-      (let* ((port (get-port server))
-             (interface (get-interface server))
-             (app (builder
-                   :session
-                   (lambda (env)
-                     (handle-http-request server env)
-                     ;; (handler-case ()
-                     ;;   (t (condition)
-                     ;;     (let* ((traceback (with-output-to-string (stream)
-                     ;;                         (trivial-backtrace:print-condition condition stream)))
-                     ;;            (condition (describe condition))
-                     ;;            (just-traceback (trivial-backtrace:backtrace-string)))
-                     ;;       (log:error "Unhandled exception" condition traceback just-traceback))
-                     ;;     '(500
-                     ;;       ("Content-Type" "text/html")
-                     ;;       ("Something went wrong!"))))
-                     ))))
-        (log:info "Starting webserver on" interface port debug)
-        
-        ;; Suppressing output to stdout, because Clack writes message
-        ;; about started server and we want to write into a log instead.
-        (with-output-to-string (*standard-output*)
-          (setf (get-handler server)
-                (clackup app
-                         :address interface
-                         :server (get-server-type server)
-                         :port port
-                         :debug debug)))))
+  (cond ((get-handler server)
+         (log:warn "Webserver already started"))
+
+        ;; Otherwise, starting a server
+        (t
+         (let* ((port (get-port server))
+                (interface (get-interface server))
+                (app (builder
+                      (make-session-middleware)
+                      (lambda (env)
+                        (handle-http-request server env)
+                        ;; Don't remember, why this code was commented
+                        ;; TODO: check how 500 errors are handled and may be remove
+                        ;;       this code, if everything is handled in some other place.
+                        ;; (handler-case ()
+                        ;;   (t (condition)
+                        ;;     (let* ((traceback (with-output-to-string (stream)
+                        ;;                         (trivial-backtrace:print-condition condition stream)))
+                        ;;            (condition (describe condition))
+                        ;;            (just-traceback (trivial-backtrace:backtrace-string)))
+                        ;;       (log:error "Unhandled exception" condition traceback just-traceback))
+                        ;;     '(500
+                        ;;       ("Content-Type" "text/html")
+                        ;;       ("Something went wrong!"))))
+                        ))))
+           (log:info "Starting webserver on" interface port debug)
+
+           ;; Suppressing output to stdout, because Clack writes message
+           ;; about started server and we want to write into a log instead.
+           (setf *clack-output* (make-string-output-stream))
+           (let ((*standard-output* *clack-output*))
+             (setf (get-handler server)
+                   (clackup app
+                            :address interface
+                            :server (get-server-type server)
+                            :port port
+                            :debug debug
+                            ;; Here we are turning off :backtrace middleware from the Lack
+                            ;; and probably all other middlewares which may become "default"
+                            ;; in future.
+                            :use-default-middlewares nil))))))
   server)
 
 
@@ -242,10 +267,6 @@ declared AUTOSTART."
 
     (reset-routes)
   
-    (unless (member :bordeaux-threads *features*)
-      (cerror "I know what I'm doing and will stubbornly continue."
-              "You're trying to start Weblocks without threading ~
-            support. Recompile your Lisp with threads enabled."))
     (if debug
         (weblocks/debug:on)
         (weblocks/debug:off))