+++ t/response.lisp
@@ -3,7 +3,8 @@
         #:rove
         #:weblocks-test/utils)
   (:import-from #:weblocks/response
-                #:make-uri))
+                #:make-uri
+                #:add-retpath-to))
 (in-package weblocks-test/response)
 
 
@@ -37,3 +38,19 @@
       (ok (equal (make-uri "../minor")
                  "http://example.com:10050/foo/minor")
           "Two dots can be used to go to the upper level of the path"))))
+
+
+(deftest add-retpath-to-test
+  (with-session
+    (with-request ("http://example.com:10050/foo/bar")
+      (ok (equal (add-retpath-to "/login")
+                 "/login?retpath=http%3A%2F%2Fexample.com%3A10050%2Ffoo%2Fbar")
+          "Retpath should be added as a single parameter.")
+      
+      (ok (equal (add-retpath-to "/login?retpath=%2Fprojects")
+                 "/login?retpath=http%3A%2F%2Fexample.com%3A10050%2Ffoo%2Fbar")
+          "Existing retpath should be overwritten.")
+      
+      (ok (equal (add-retpath-to "/login?code=100500&retpath=%2Fprojects")
+                 "/login?code=100500&retpath=http%3A%2F%2Fexample.com%3A10050%2Ffoo%2Fbar")
+          "Existing parameters should be leaved as is."))))