+++ ChangeLog.rst
@@ -2,6 +2,106 @@
  ChangeLog
 ===========
 
+0.39.1 (2020-01-20)
+===================
+
+Fixed
+-----
+
+* Fixed processing of the value ``x-forwarded-port``. Previously, when Weblocks
+  was used along with lastest (after the 2019-09-07) ``Woo`` server, this header
+  lead to the 500 error. Because behavior of the ``Woo`` was changed:
+
+  https://github.com/fukamachi/woo/issues/84
+
+  Now Woo does not parses numberic headers and Weblocks has to do it itself.
+
+0.39.0 (2019-09-16)
+===================
+
+Changes
+-------
+
+* A new macro ``weblocks/routes:defroute`` was added.
+
+  It defines a handler for a given route. By default route should return
+  a serialized JSON:
+
+  .. code:: common-lisp
+            
+     (defroute (app /api/data)
+         "{\"my-data\": [1, 2, 3]}")
+
+  but you can redefine the content type:
+
+  .. code:: common-lisp
+ 
+     (defroute (app /api/data :content-type "application/xml")
+         "<my-data><item>1</item><item>2</item></my-data>")
+
+  each route is associate with application class and these routes
+  are added to the mapper when you call a ``start`` method.
+
+
+0.38.1 (2019-08-02)
+===================
+
+* A small fix to make it work with ``weblocks-websocket``.
+
+0.38.0 (2019-07-07)
+===================
+
+* Unhandled error processing was fixed. There were numerous problems with error processing:
+  * Previosly server stopped to work because clack tried to write into a closed stream.
+  * Also, before this fix, error page didn't showed propertly.
+  * Any error in the route lead to a server stop because there wasn't any handler at all.
+    Now they are handled with ``on-error`` method as errors in block/page rendering.
+
+  All of them are fixed now.
+
+0.37.0 (2019-06-01)
+===================
+
+* Removed check if ``:bordeaux-threads`` is in the ``*features*``, because this does not work
+  in runtime in precompiled binary.
+
+0.36.0(2019-05-03)
+===================
+
+* Added function ``weblocks/session:expire``, which can be used to tell Weblocks to delete current session
+  after request processing.
+* Also, now session middleware is created explicitly along with memory store. This made possible to add
+  another two functions to gather some statistics about number of sessions:
+
+  * ``weblocks/session:get-number-of-sessions``;
+  * ``weblocks/session:get-number-of-anonymous-sessions``;
+  
+* The function ``weblocks/widgets/base:get-html-tag`` was fixed to
+  render tables correctly. If the widget's parent is a ``:table``
+  element, then this function returns ``:tr`` (instead of a
+  ``:div``). If it is ``:tr``, it returns ``:td``, and ``:div``
+  otherwise.
+* A new setting ``weblocks/html:*pretty-html*`` was created, to further control pretty printing of html
+  both in code and in tests.
+
+0.35.3 (2019-03-31)
+===================
+
+* Added a ``weblocks/response:add-retpath-to`` function, useful to add a
+  ``retpath`` GET parameter to the URL.
+
+  For example, calling::
+
+    (add-retpath-to "/login" :retpath
+    "http://example.com:10050/foo/bar")
+
+  Will return::
+
+    "/login?retpath=http%3A%2F%2Fexample.com%3A10050%2Ffoo%2Fbar"
+
+  Argument ``:retpath`` is optional. By default, function
+  will take an URL of the current web page.
+
 0.35.2 (2019-03-21)
 ===================
 
@@ -104,7 +204,7 @@ Improvements
   ``log4cl-json`` library. To turn it on, just do:
 
   .. code:: common-lisp
-            
+
      (log4cl-json:setup)
 
 Fixes
@@ -121,11 +221,11 @@ Fixes
   For example:
 
   .. code:: common-lisp
-         
+
 
      (weblocks/hooks:on-application-hook-handle-request
        connect-to-database ()
-       
+
        (let ((success nil))
          (unwind-protect (progn (setup-transaction)
                                 (weblocks/hooks:call-next-hook)
@@ -216,7 +316,7 @@ Fixes
   crashes with error::
 
     Assertion failed: (("libev: a signal must not be attached to two different loops", !signals [w->signum - 1].loop || signals [w->signum - 1].loop == loop)), function ev_signal_start, file ev.c, line 4082
-  
+
 
 0.30.0 (2018-05-19)
 ===================
@@ -515,7 +615,7 @@ And macro for defining a special readers for them was removed as well:
 
 Also, these arguments to ``defapp`` was removed:
 ``:ignore-default-dependencies``, ``:dependencies``
-  
+
 Function ``update-thread-status`` and method ``webapp-update-thread-status``
 were removed.
 
@@ -804,7 +904,7 @@ can be useful when embedding links into emails, for example.
 * Now we are trying to call action only if action's name was given.
 * Old method ``weblocks:handle-client-request ((app weblocks-webapp))``
   was removed. Look at it's newer version in ``weblocks.request-handler``.
-  
+
 
 0.18.0 (2017-11-12)
 ===================
@@ -821,7 +921,7 @@ can be useful when embedding links into emails, for example.
 * Call to ``prepare-hooks`` was moved from ``weblocks.request-handler:handle-client-request``
   to the the weblocks.server:handler-request, to fix session hooks processing when
   ``:process-request`` hook is called.
-  
+
 0.17.2 (2017-11-11)
 ===================
 
@@ -855,7 +955,7 @@ can be useful when embedding links into emails, for example.
   executed in same order as they were added. If you want some code to be
   executed before widget update, just execute ``send-code`` before
   ``weblocks.widget:update``.
-  
+
 0.16.0 (2017-11-04)
 ===================
 
@@ -902,10 +1002,10 @@ can be useful when embedding links into emails, for example.
 
   Also, a new macro ``catch-hooks`` was added to check if some
   hooks were called during a unittest.
-  
+
 * Now weblocks does not open a new tab or window on 500 error
   during an action execution.
-  
+
 0.14.4 (2017-10-07)
 ===================
 
@@ -919,10 +1019,10 @@ can be useful when embedding links into emails, for example.
 
 * Default method of ``render-page`` was fixed to really wrap
   page with ``<html>...`` block.
-  
+
 * Fixed a way how weblocks.debug:*latest-session* is
   processed.
-  
+
 * Function ``weblocks.request:remove-request-header`` now
   returns a new instance of request object and does not modify the
   original request. This fixes issue in ``weblocks-websocket``.
@@ -978,7 +1078,7 @@ can be useful when embedding links into emails, for example.
 
 0.13.10 (2017-09-06)
 ====================
-  
+
 Changes in weblocks.request-hooks:
 ----------------------------------
 
@@ -989,7 +1089,7 @@ Changes in weblocks.request-hooks:
   should be used like:
 
   .. code-block:: lisp
-       
+
      (weblocks.hooks:add-session-hook
               :some-hook
               my-beautiful-callback (param)
@@ -1003,7 +1103,7 @@ Changes in weblocks.request-hooks:
      (weblocks.hooks:call :some-hook
            first-param
            second-param)
-           
+
 0.13.10 (2017-09-06)
 ====================
 
@@ -1018,7 +1118,7 @@ Changes in weblocks.request-hooks:
   Also, ``get-cross-origin`` and ``:cross-origin`` were removed to
   ``get-crossorigin`` and ``:crossorigin``, to conform the html
   attibute's spelling.
-  
+
 0.13.8 (2017-09-02)
 ===================
 
@@ -1034,7 +1134,7 @@ Changes in weblocks.request-hooks:
 
 * Reformatted documentation. Started to `keep a changelog
   <http://keepachangelog.com/>`_.
-  
+
 0.13.7 (2017-04-15)
 ===================